feat: Implement simple ptr-sli

+++ dh/include/dh/claim/assert_static.h
@@ -42,19 +42,19 @@ extern "C" {
 #else
 /* Older versions - emulate static assert */
 
-#include "../core/pp.h"
-#include "../core/prim.h"
+#include "dh/core/pp.h"
+#include "dh/core/prim.h"
 
 #ifdef __COUNTER__
 #define IMPL_claim_assert_static(_Expr) \
-    typedef i8 pp_cat(claim_assert_static_, __COUNTER__)[(_Expr) ? 1 : -1]
+    typedef int pp_cat(claim_assert_static_, __COUNTER__)[(_Expr) ? 1 : -1]
 #define IMPL_claim_assert_static_msg(_Expr, _msg) \
-    typedef i8 pp_cat(claim_assert_static_msg_, __COUNTER__)[(_Expr) ? 1 : -1]
+    typedef int pp_cat(claim_assert_static_msg_, __COUNTER__)[(_Expr) ? 1 : -1]
 #else
 #define IMPL_claim_assert_static(_Expr) \
-    typedef i8 pp_cat(claim_assert_static_, __LINE__)[(_Expr) ? 1 : -1]
+    typedef int pp_cat(claim_assert_static_, __LINE__)[(_Expr) ? 1 : -1]
 #define IMPL_claim_assert_static_msg(_Expr, _msg) \
-    typedef i8 pp_cat(claim_assert_static_msg_, __LINE__)[(_Expr) ? 1 : -1]
+    typedef int pp_cat(claim_assert_static_msg_, __LINE__)[(_Expr) ? 1 : -1]
 #endif
 
 #endif
